chore: dedup WaitForString() into testutils (#712)
Dedup the `WaitForString()` helper function that is used in `root_test` and db integration tests.
